test: enable PreferencesManager-integ-test
